refactor: modularize Notebook component logic into separate interaction, template, and utility modules
