+++ src/ravioli/frontend/src/components/Notebook.ts
@@ -1,327 +1,9 @@
 import { store } from '../store';
 import { api } from '../services/api';
-import MarkdownIt from 'markdown-it';
-import Chart from 'chart.js/auto';
 import { format } from 'date-fns';
-
-import hljs from 'highlight.js';
-import 'highlight.js/styles/github-dark.css';
-
-const md = new MarkdownIt({
-  html: false,
-  linkify: true,
-  typographer: true,
-  highlight: function (str: string, lang: string) {
-    if (lang && hljs.getLanguage(lang)) {
-      try {
-        return '<pre class="hljs"><code>' +
-               hljs.highlight(str, { language: lang, ignoreIllegals: true }).value +
-               '</code></pre>';
-      } catch (__) {}
-    }
-
-    return ''; // use external default escaping
-  }
-});
-
-type CellType = 'python' | 'sql' | 'chat' | 'markdown';
-
-function parseCellType(value: string | null): CellType | null {
-  return value === 'python' || value === 'sql' || value === 'chat' || value === 'markdown' ? value : null;
-}
-
-function parseAfterLogId(value: string | null): string | null {
-  if (!value) return null;
-  if (value === '__first__') return value;
-  return /^[A-Za-z0-9_-]+$/.test(value) ? value : null;
-}
-
-export function highlightSQL(code: string): string {
-  if (!code) return '';
-  
-  let html = code
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;');
-    
-  const sqlRegex = /(--[^\n]*)|((['"])(?:[^\\]|\\.)*?\3)|(\b\d+\b)|(\b(?:SELECT|FROM|WHERE|AND|OR|NOT|LIMIT|OFFSET|INSERT|UPDATE|DELETE|CREATE|DROP|ALTER|TABLE|JOIN|INNER|LEFT|RIGHT|OUTER|ON|GROUP|BY|ORDER|HAVING|AS|IN|IS|NULL|LIKE|ILIKE|WITH|UNION|ALL|CASE|WHEN|THEN|ELSE|END|COUNT|SUM|AVG|MIN|MAX|CAST|COALESCE|DISTINCT)\b)/gi;
-
-  return html.replace(sqlRegex, (match, comment, stringVal, quote, numberVal, keyword) => {
-    if (comment) {
-      return `<span class="text-neutral-500 italic">${comment}</span>`;
-    }
-    if (stringVal) {
-      return `<span class="text-emerald-400">${stringVal}</span>`;
-    }
-    if (numberVal) {
-      return `<span class="text-amber-400">${numberVal}</span>`;
-    }
-    if (keyword) {
-      return `<span class="text-sky-400 font-bold">${keyword}</span>`;
-    }
-    return match;
-  });
-}
-
-export function highlightPython(code: string): string {
-  if (!code) return '';
-
-  let html = code
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;');
-
-  // Order: comments, strings, numbers, decorators, keywords, builtins
-  const pyRegex = /(#[^\n]*)|("""[\s\S]*?"""|'''[\s\S]*?'''|"(?:[^\\"]|\\.)*"|'(?:[^\\']|\\.)*')|((?<!\w)\d+\.?\d*(?:[eE][+-]?\d+)?(?![a-zA-Z_]))|(@[\w.]+)|(\b(?:False|None|True|and|as|assert|async|await|break|class|continue|def|del|elif|else|except|finally|for|from|global|if|import|in|is|lambda|nonlocal|not|or|pass|raise|return|try|while|with|yield)\b)|(\b(?:abs|all|any|bool|bytes|callable|chr|dict|dir|divmod|enumerate|eval|exec|filter|float|format|getattr|globals|hasattr|hash|help|hex|id|input|int|isinstance|issubclass|iter|len|list|locals|map|max|memoryview|min|next|object|oct|open|ord|pow|print|property|range|repr|reversed|round|set|setattr|slice|sorted|staticmethod|str|sum|super|tuple|type|vars|zip|pd|np|plt|con|df|execute|fetchone|fetchall|fetchdf|pl)\b)/gm;
-
-  return html.replace(pyRegex, (match, comment, stringVal, numberVal, decorator, keyword, builtin) => {
-    if (comment)   return `<span class="text-neutral-500 italic">${comment}</span>`;
-    if (stringVal) return `<span class="text-amber-300">${stringVal}</span>`;
-    if (numberVal) return `<span class="text-violet-400">${numberVal}</span>`;
-    if (decorator) return `<span class="text-pink-400">${decorator}</span>`;
-    if (keyword)   return `<span class="text-sky-400 font-bold">${keyword}</span>`;
-    if (builtin)   return `<span class="text-emerald-300">${builtin}</span>`;
-    return match;
-  });
-}
-
-function toggleComment(textarea: HTMLTextAreaElement) {
-  const value = textarea.value;
-  const start = textarea.selectionStart;
-  const end = textarea.selectionEnd;
-  
-  const before = value.substring(0, start);
-  const after = value.substring(start);
-  
-  const lineStartIdx = before.lastIndexOf('\n') + 1;
-  const nextNewline = after.indexOf('\n');
-  const lineEndIdx = nextNewline !== -1 ? start + nextNewline : value.length;
-  
-  const currentLine = value.substring(lineStartIdx, lineEndIdx);
-  
-  let newLine: string;
-  let offset: number;
-  
-  if (/^\s*--/.test(currentLine)) {
-    newLine = currentLine.replace(/^(\s*)--\s?/, '$1');
-    offset = newLine.length - currentLine.length;
-  } else {
-    newLine = currentLine.replace(/^(\s*)/, '$1-- ');
-    offset = newLine.length - currentLine.length;
-  }
-  
-  textarea.value = value.substring(0, lineStartIdx) + newLine + value.substring(lineEndIdx);
-  
-  textarea.selectionStart = start + (start >= lineStartIdx + (currentLine.match(/^\s*/)?.[0].length || 0) ? offset : 0);
-  textarea.selectionEnd = end + (end >= lineStartIdx + (currentLine.match(/^\s*/)?.[0].length || 0) ? offset : 0);
-}
-
-function updateLineNumbers(textarea: HTMLTextAreaElement) {
-  const idStr = textarea.id.replace('cell-input-', '');
-  const parent = textarea.closest('.glass-panel');
-  if (parent) {
-    const gutter = parent.querySelector(`#cell-gutter-${idStr}`) as HTMLElement;
-    if (gutter) {
-      const lines = textarea.value.split('\n');
-      const lineNumbers = lines.map((_, i) => i + 1).join('\n');
-      gutter.textContent = lineNumbers;
-      gutter.scrollTop = textarea.scrollTop;
-    }
-
-    const highlight = parent.querySelector(`#cell-highlight-${idStr}`) as HTMLElement;
-    if (highlight) {
-      const tool = highlight.getAttribute('data-tool');
-      if (tool === 'sql') {
-        highlight.innerHTML = highlightSQL(textarea.value);
-      } else if (tool === 'python') {
-        highlight.innerHTML = highlightPython(textarea.value);
-      } else {
-        highlight.textContent = textarea.value;
-      }
-      highlight.scrollTop = textarea.scrollTop;
-      highlight.scrollLeft = textarea.scrollLeft;
-    }
-  }
-}
-
-function updateLineGutterScroll(textarea: HTMLTextAreaElement) {
-  const idStr = textarea.id.replace('cell-input-', '');
-  const parent = textarea.closest('.glass-panel');
-  if (parent) {
-    const gutter = parent.querySelector(`#cell-gutter-${idStr}`) as HTMLElement;
-    if (gutter) {
-      gutter.scrollTop = textarea.scrollTop;
-    }
-
-    const highlight = parent.querySelector(`#cell-highlight-${idStr}`) as HTMLElement;
-    if (highlight) {
-      highlight.scrollTop = textarea.scrollTop;
-      highlight.scrollLeft = textarea.scrollLeft;
-    }
-  }
-}
-
-function renderMarkdown(content: string) {
-  if (!content) return '';
-  
-  // Transform GitHub style alerts: > [!TYPE]
-  // This regex matches the blockquote with alert marker
-  let transformed = content.replace(/^> \[!(IMPORTANT|NOTE|TIP|WARNING|CAUTION)\]\n((?:>.*\n?)+)/gm, (_match, type, body) => {
-    const lowerType = type.toLowerCase();
-    const icon = type === 'IMPORTANT' ? 'priority_high' : 'info';
-    // Remove the leading '>' from each line of the body
-    const cleanBody = body.split('\n').map((line: string) => line.replace(/^>\s?/, '')).join('\n');
-    return `
-<div class="markdown-alert markdown-alert-${lowerType}">
-  <div class="markdown-alert-title">
-    <span class="material-symbols-outlined text-sm" data-icon="${icon}">${icon}</span>
-    <span>${type}</span>
-  </div>
-  <div class="markdown-alert-content">
-    ${md.render(cleanBody.trim())}
-  </div>
-</div>
-`;
-  });
-
-  return md.render(transformed);
-}
-
-function renderChart(canvasId: string, vizData: any) {
-  const ctx = document.getElementById(canvasId) as HTMLCanvasElement;
-  if (!ctx) return;
-
-  new Chart(ctx, {
-    type: vizData.chart_type,
-    data: vizData.data,
-    options: {
-      responsive: true,
-      maintainAspectRatio: false,
-      plugins: {
-        legend: {
-          display: true,
-          position: 'bottom',
-          labels: {
-            color: '#94a3b8',
-            font: { family: 'Inter', size: 10 }
-          }
-        },
-        title: {
-          display: true,
-          text: vizData.title,
-          color: '#f8fafc',
-          font: { family: 'Inter', size: 14, weight: 'bold' }
-        }
-      },
-      scales: {
-        x: {
-          grid: { color: 'rgba(255, 255, 255, 0.05)' },
-          ticks: { color: '#64748b', font: { size: 10 } }
-        },
-        y: {
-          grid: { color: 'rgba(255, 255, 255, 0.05)' },
-          ticks: { color: '#64748b', font: { size: 10 } }
-        }
-      }
-    }
-  });
-}
-
-function renderRichOutput(output: any): string {
-  if (!output) return '';
-  
-  if (output.type === 'stream') {
-    const isStderr = output.name === 'stderr';
-    const textColor = isStderr ? 'text-rose-400' : 'text-neutral-300';
-    return `
-      <div class="font-mono text-xs ${textColor} bg-surface-container-lowest/40 p-3 rounded-lg border border-outline-variant/5 overflow-x-auto whitespace-pre-wrap leading-relaxed">
-        ${output.text}
-      </div>
-    `;
-  }
-  
-  if (output.type === 'execute_result' || output.type === 'display_data') {
-    const data = output.data || {};
-    if (data['image/png']) {
-      return `
-        <div class="mt-2 glass-panel p-4 rounded-2xl border-outline-variant/10 bg-surface-container-lowest/30 flex justify-center overflow-hidden">
-          <img src="data:image/png;base64,${data['image/png'].trim()}" alt="Plot Output" class="max-w-full h-auto rounded-lg" />
-        </div>
-      `;
-    }
-    if (data['text/html']) {
-      return `
-        <div class="mt-2 overflow-x-auto custom-scrollbar bg-surface-container-lowest/50 rounded-2xl border border-outline-variant/10 max-h-96">
-          <div class="prose prose-invert max-w-none text-xs font-mono p-4 table-rendering-wrapper">
-            ${data['text/html']}
-          </div>
-        </div>
-      `;
-    }
-    if (data['text/plain']) {
-      return `
-        <div class="font-mono text-xs text-emerald-300 bg-surface-container-lowest/40 p-3 rounded-lg border border-outline-variant/5 overflow-x-auto whitespace-pre leading-relaxed">
-          ${data['text/plain']}
-        </div>
-      `;
-    }
-  }
-  
-  if (output.type === 'error') {
-    const traceback = output.traceback || [];
-    const tracebackHtml = traceback.length > 0 
-      ? `<div class="mt-2 p-3 font-mono text-[11px] bg-red-950/20 text-rose-300/80 rounded-lg border border-red-500/10 overflow-x-auto whitespace-pre leading-relaxed">${traceback.join('\n')}</div>`
-      : '';
-    return `
-      <div class="mt-2 glass-panel p-4 rounded-2xl border-rose-500/20 bg-rose-500/5 text-rose-200">
-        <div class="flex items-center gap-2 text-rose-400 font-bold text-sm">
-          <span class="material-symbols-outlined text-lg">error</span>
-          <span>${output.ename || 'Execution Error'}: ${output.evalue || 'Something went wrong'}</span>
-        </div>
-        ${tracebackHtml}
-      </div>
-    `;
-  }
-  
-  if (output.type === 'table') {
-    const rows = output.data || [];
-    if (rows.length === 0) {
-      return `
-        <div class="flex flex-col items-center justify-center p-6 text-outline bg-surface-container-lowest/30 rounded-2xl border border-outline-variant/5">
-          <span class="material-symbols-outlined text-3xl mb-1 opacity-40">draft</span>
-          <span class="text-xs">No records returned</span>
-        </div>
-      `;
-    }
-    
-    const keys = Object.keys(rows[0]);
-    return `
-      <div class="mt-2 overflow-x-auto custom-scrollbar bg-surface-container-lowest/50 rounded-2xl border border-outline-variant/10 max-h-96">
-        <table class="w-full text-left border-collapse text-xs font-mono">
-          <thead class="bg-surface-container-highest text-xs uppercase tracking-wider text-primary sticky top-0 shadow-sm z-10">
-            <tr>
-              ${keys.map(k => `<th class="px-4 py-2.5 font-medium whitespace-nowrap border-b border-primary/20">${k}</th>`).join('')}
-            </tr>
-          </thead>
-          <tbody class="divide-y divide-outline-variant/10">
-            ${rows.map((row: any) => `
-              <tr class="hover:bg-surface-container-low/70 transition-colors">
-                ${keys.map(k => {
-                  let val = row[k];
-                  if (val === null || val === undefined) return '<td class="px-4 py-2 text-outline/40 italic text-[11px]">null</td>';
-                  if (typeof val === 'object') val = JSON.stringify(val);
-                  return `<td class="px-4 py-2 text-on-surface-variant truncate max-w-sm" title="${val}">${val}</td>`;
-                }).join('')}
-              </tr>
-            `).join('')}
-          </tbody>
-        </table>
-      </div>
-    `;
-  }
-  
-  return '';
-}
+import { highlightSQL, highlightPython } from './notebook/utils';
+import { renderMarkdown, renderChart, renderRichOutput } from './notebook/templates';
+import { bindInteractions } from './notebook/interactions';
 
 let lastLogsJson = '';
 const executingCells = new Set<string>();
@@ -523,7 +205,7 @@ export function updateNotebookUI(container: HTMLElement, isInitial = false) {
       </div>
     `;
     
-    bindInteractions(container);
+    bindInteractions(container, updateNotebookUI);
   }
 
   // Update Status & Compact Metadata Header
@@ -946,768 +628,3 @@ export function updateNotebookUI(container: HTMLElement, isInitial = false) {
   }
 }
 
-function bindInteractions(container: HTMLElement) {
-  const activeId = store.getActiveAnalysisId();
-
-  // Double click to edit cell static view
-  container.addEventListener('dblclick', (e) => {
-    const staticView = (e.target as HTMLElement).closest('.cell-static-view') as HTMLElement;
-    if (staticView) {
-      const editBtn = staticView.querySelector('.btn-edit-cell') as HTMLElement;
-      editBtn?.click();
-    }
-  });
-
-  // Real-time LOC line counting & Gutter update
-  container.addEventListener('input', (e) => {
-    const textarea = e.target as HTMLTextAreaElement;
-    if (textarea && textarea.id && textarea.id.startsWith('cell-input-')) {
-      const idxStr = textarea.id.replace('cell-input-', '');
-      const lineCount = textarea.value.split('\n').length;
-      const label = container.querySelector(`#cell-loc-${idxStr}`);
-      if (label) {
-        label.textContent = `${lineCount} ${lineCount === 1 ? 'line' : 'lines'}`;
-      }
-      // Update gutter line numbers & highlight backdrop
-      updateLineNumbers(textarea);
-    }
-  });
-
-  // Attach scroll & initial line counts on focus
-  container.addEventListener('focusin', (e) => {
-    const textarea = e.target as HTMLTextAreaElement;
-    if (textarea && textarea.id && textarea.id.startsWith('cell-input-')) {
-      // Sync line numbers immediately
-      updateLineNumbers(textarea);
-      
-      // Bind scroll sync if not already bound
-      if (!textarea.dataset.hasScrollListener) {
-        textarea.dataset.hasScrollListener = 'true';
-        textarea.addEventListener('scroll', () => {
-          updateLineGutterScroll(textarea);
-        });
-      }
-    }
-  });
-
-  // Hotkeys: Ctrl+Enter to execute, Cmd+/ or Ctrl+/ to toggle comment
-  container.addEventListener('keydown', (e) => {
-    const textarea = e.target as HTMLTextAreaElement;
-    if (textarea && textarea.id && textarea.id.startsWith('cell-input-')) {
-      const isCmdOrCtrl = e.metaKey || e.ctrlKey;
-
-      // Ctrl + Enter to run
-      if (e.key === 'Enter' && e.ctrlKey) {
-        e.preventDefault();
-        const card = textarea.closest('.glass-panel');
-        if (card) {
-          const runBtn = card.querySelector('.btn-rerun-cell, .btn-execute-new-cell') as HTMLElement;
-          runBtn?.click();
-        }
-      }
-
-      // Cmd + / or Ctrl + / to toggle comment
-      if ((e.key === '/' || e.code === 'Slash') && isCmdOrCtrl) {
-        e.preventDefault();
-        toggleComment(textarea);
-        // Trigger input event to update line counts and autosize
-        textarea.dispatchEvent(new Event('input', { bubbles: true }));
-      }
-    }
-  });
-
-  function createCellEditBlock(
-    type: 'python' | 'sql' | 'chat' | 'markdown',
-    afterLogId: string | null,
-    targetElement: HTMLElement,
-    toolbarNode?: HTMLElement
-  ) {
-    const tempId = 'temp-' + Date.now();
-    const newCell = document.createElement('div');
-    newCell.className = 'glass-panel p-6 rounded-3xl space-y-6 bg-surface-container-low/30 border-primary/40 relative overflow-hidden group animate-in fade-in slide-in-from-top-4 duration-500 shadow-xl shadow-primary/5 new-cell-block';
-    
-    let icon = "auto_awesome";
-    let color = "text-secondary";
-    let placeholder = "Enter AI instruction...";
-    
-    if (type === "sql") {
-       icon = "database";
-       color = "text-primary";
-       placeholder = "SELECT * FROM ...";
-    } else if (type === "python") {
-       icon = "code";
-       color = "text-emerald-400";
-       placeholder = "# e.g. df = con.table('my_schema.my_table')  →  df.head()";
-    } else if (type === "markdown") {
-       icon = "article";
-       color = "text-indigo-400";
-       placeholder = "Write notes or analysis findings...";
-    }
-
-    const isMarkdown = type === "markdown";
-    const borderFocusClass = type === "markdown" ? "indigo-400" : (type === "sql" ? "primary" : (type === "python" ? "emerald-400" : "secondary"));
-
-    newCell.innerHTML = `
-       <div class="flex items-start gap-4">
-          ${isMarkdown ? '' : `
-          <div class="font-mono text-xs font-bold ${color}/80 pt-2.5 select-none w-14 text-right shrink-0 flex items-center justify-end gap-1">
-             <span class="material-symbols-outlined text-[14px]">${icon}</span>
-             In [*]:
-          </div>
-          `}
-          
-          <div class="flex-1 w-full" id="cell-edit-${tempId}">
-             <div class="glass-panel p-1.5 rounded-xl group focus-within:border-${borderFocusClass}/30 transition-all duration-300 shadow-lg shadow-${borderFocusClass}/5 bg-surface-container-lowest/80 border-outline-variant/10">
-                <div class="flex flex-col w-full">
-                  <div class="flex items-stretch gap-3 px-3">
-                    ${(type === 'sql' || type === 'python') ? `
-                    <div id="cell-gutter-${tempId}" class="w-8 select-none text-right font-mono text-sm leading-relaxed text-outline/30 pr-2 border-r border-outline-variant/10 whitespace-pre overflow-hidden pt-0 pointer-events-none">1</div>
-                    ` : ''}
-                    <div class="flex-1 min-w-0 py-0.5 relative">
-                      ${type === 'sql' || type === 'python' ? `
-                      <!-- Highlight Backdrop -->
-                      <div id="cell-highlight-${tempId}" data-tool="${type}" class="absolute inset-0 w-full max-w-full bg-transparent text-on-surface py-0.5 px-0 text-sm font-mono leading-relaxed whitespace-pre overflow-hidden pointer-events-none border border-transparent custom-scrollbar select-none"></div>
-                      ` : ''}
-                      <textarea id="cell-input-${tempId}" class="w-full max-w-full bg-transparent border border-transparent ${(type === 'sql' || type === 'python') ? 'text-transparent caret-white whitespace-pre overflow-x-auto' : 'text-on-surface whitespace-pre-wrap break-words'} outline-none focus:outline-none focus:ring-0 resize-none py-0.5 px-0 text-sm font-mono leading-relaxed custom-scrollbar placeholder-outline-variant relative z-10" rows="${isMarkdown ? 3 : 2}" placeholder="${placeholder}"></textarea>
-                    </div>
-                    <div class="flex items-start gap-1.5 shrink-0 pt-0.5">
-                      <button class="w-7 h-7 rounded-full bg-surface-container-highest text-outline flex items-center justify-center hover:bg-error/20 hover:text-error transition-colors btn-cancel-insert" title="Cancel">
-                        <span class="material-symbols-outlined text-[14px]">close</span>
-                      </button>
-                      <button class="w-7 h-7 rounded-full bg-primary text-on-primary flex items-center justify-center hover:scale-110 transition-transform btn-execute-new-cell shadow-md shadow-primary/20" data-type="${type}" data-after="${afterLogId || '__first__'}" data-temp-id="${tempId}" title="Run Cell">
-                        <span class="material-symbols-outlined text-[14px]">${isMarkdown ? 'done' : 'play_arrow'}</span>
-                      </button>
-                    </div>
-                  </div>
-                  ${type === 'sql' ? `
-                  <div class="px-3 pb-1.5 flex items-center justify-between text-[10px] font-mono text-outline select-none border-t border-outline-variant/5 pt-1.5 mt-1.5">
-                    <span>SQL Mode</span>
-                    <span id="cell-loc-${tempId}">1 line</span>
-                  </div>
-                  ` : ''}
-                </div>
-             </div>
-          </div>
-       </div>
-    `;
-
-    if (toolbarNode) {
-      toolbarNode.classList.add('hidden');
-    }
-
-    targetElement.insertAdjacentElement('afterend', newCell);
-    
-    // Auto-focus new textarea
-    const txt = newCell.querySelector(`#cell-input-${tempId}`) as HTMLTextAreaElement;
-    if (txt) {
-       txt.focus();
-       // Shift+Enter to run
-       txt.addEventListener('keydown', (ev) => {
-         if (ev.key === 'Enter' && ev.shiftKey) {
-           ev.preventDefault();
-           newCell.querySelector('.btn-execute-new-cell')?.dispatchEvent(new Event('click', { bubbles: true }));
-         }
-       });
-    }
-
-    // Handle Cancel Insert
-    const cancelBtn = newCell.querySelector('.btn-cancel-insert');
-    cancelBtn?.addEventListener('click', () => {
-       newCell.remove();
-       if (toolbarNode) {
-          toolbarNode.classList.remove('hidden');
-       }
-       // If no cells are left in the container, restore welcome screen
-       const cellContainer = container.querySelector('#cell-container');
-       if (cellContainer && !cellContainer.querySelector('.glass-panel')) {
-          updateNotebookUI(container);
-       }
-    });
-  }
-
-  const input = container.querySelector('#cell-input') as HTMLTextAreaElement;
-  const btn = container.querySelector('#btn-execute');
-  const btnMagic = container.querySelector('#btn-magic');
-  const magicPopover = container.querySelector('#magic-popover');
-  const suggestionsList = container.querySelector('#magic-suggestions-list');
-
-  // Magic Suggestions
-  btnMagic?.addEventListener('click', async (e) => {
-    e.stopPropagation();
-    if (!activeId) return;
-
-    if (magicPopover?.classList.contains('hidden')) {
-      magicPopover.classList.remove('hidden');
-      if (suggestionsList) {
-        suggestionsList.innerHTML = `<div class="py-4 flex flex-col items-center gap-2 opacity-50"><span class="material-symbols-outlined animate-spin text-lg" data-icon="progress_activity">progress_activity</span></div>`;
-        try {
-          const prompts = await api.getSuggestedPrompts(activeId);
-          suggestionsList.innerHTML = prompts.map(p => `
-            <button class="magic-suggestion-item w-full text-left px-3 py-2 text-xs font-body-sm text-on-surface-variant hover:bg-primary/10 hover:text-primary rounded-lg" data-prompt="${p.replace(/"/g, '&quot;')}">${p}</button>
-          `).join('');
-          suggestionsList.querySelectorAll('.magic-suggestion-item').forEach(item => {
-            item.addEventListener('click', () => {
-              const prompt = item.getAttribute('data-prompt');
-              if (prompt && input) {
-                input.value = prompt;
-                input.dispatchEvent(new Event('input'));
-                magicPopover.classList.add('hidden');
-                input.focus();
-              }
-            });
-          });
-        } catch (err) { suggestionsList.innerHTML = `<div class="text-[10px] text-error p-2">Error</div>`; }
-      }
-    } else { magicPopover?.classList.add('hidden'); }
-  });
-
-  document.addEventListener('click', () => magicPopover?.classList.add('hidden'));
-
-  // Footer "Add Cell" bar
-  container.querySelector('#add-cell-bar')?.addEventListener('click', (e) => {
-    const addBtn = (e.target as HTMLElement).closest('.btn-add-cell') as HTMLElement;
-    if (!addBtn) return;
-    const type = parseCellType(addBtn.getAttribute('data-type'));
-    if (!type) return;
-    
-    const cellContainer = container.querySelector('#cell-container');
-    if (!cellContainer) return;
-
-    // Remove the welcome screen if it is there
-    const welcome = cellContainer.querySelector('#notebook-welcome');
-    if (welcome) welcome.remove();
-
-    // Get the last log in the store to anchor the new cell after it
-    const logs = store.getLogs();
-    const lastLog = logs[logs.length - 1];
-    const afterLogId = lastLog?.id ?? null;
-
-    // Append cell at the end of the container
-    const children = Array.from(cellContainer.children);
-    const lastChild = children[children.length - 1] as HTMLElement;
-
-    if (lastChild) {
-      createCellEditBlock(type, afterLogId, lastChild);
-    } else {
-      const dummy = document.createElement('div');
-      dummy.className = 'hidden';
-      cellContainer.appendChild(dummy);
-      createCellEditBlock(type, afterLogId, dummy);
-    }
-
-    cellContainer.scrollTop = cellContainer.scrollHeight;
-  });
-
-  // Delegated events for In-place Cell Editing
-  container.addEventListener('click', async (e) => {
-    const target = e.target as HTMLElement;
-
-    async function runCell(idx: string, logId: string, toolName: string, question: string, cellBody: HTMLElement) {
-      executingCells.add(logId);
-      container.querySelector(`#cell-edit-${idx}`)?.classList.add('hidden');
-      const staticView = container.querySelector(`#cell-static-${idx}`);
-      if (staticView) {
-        staticView.classList.remove('hidden');
-        staticView.innerHTML = '';
-        const questionEl = document.createElement('div');
-        questionEl.className = 'pr-8 whitespace-pre-wrap';
-        if (toolName === 'sql') {
-          questionEl.innerHTML = highlightSQL(question);
-        } else {
-          questionEl.textContent = question;
-        }
-        staticView.appendChild(questionEl);
-      }
-      
-      const divider = container.querySelector(`#cell-divider-${idx}`);
-      if (divider) divider.classList.remove('opacity-0');
-      
-      const outGutter = cellBody.querySelector('.text-secondary\\/50') as HTMLElement;
-      if (outGutter) {
-         outGutter.innerHTML = `<span class="material-symbols-outlined text-[10px] animate-spin" data-icon="progress_activity">progress_activity</span><span>Out [*]:</span>`;
-         outGutter.classList.add('flex', 'items-center', 'justify-end', 'gap-1');
-      }
-      
-      const outBody = outGutter?.nextElementSibling;
-      if (outBody) {
-        const streamingDiv = document.createElement('div');
-        streamingDiv.className = 'prose prose-invert max-w-none text-on-surface-variant leading-relaxed font-body-lg animate-pulse';
-        streamingDiv.id = `streaming-content-${idx}`;
-
-        const cursor = document.createElement('span');
-        cursor.className = 'inline-block w-1 h-4 bg-primary animate-pulse';
-        streamingDiv.appendChild(cursor);
-
-        outBody.replaceChildren(streamingDiv);
-      }
-      
-      const streamingContent = outBody?.querySelector(`#streaming-content-${idx}`);
-
-      if (toolName === 'sql' || toolName === 'python') {
-        try {
-          if (toolName === 'sql') {
-            await api.executeSql(activeId!, question, logId, null);
-          } else {
-            await api.executePython(activeId!, question, logId, null);
-          }
-
-          if (outGutter) {
-             outGutter.textContent = `Out [${idx}]:`;
-             outGutter.classList.remove('flex', 'items-center', 'justify-end', 'gap-1');
-          }
-          
-          const newLogs = await api.listLogs(activeId!);
-          lastLogsJson = '';
-          executingCells.delete(logId);
-          store.setLogs(newLogs);
-        } catch (e) {
-           console.error(e);
-           if (outGutter) {
-              outGutter.textContent = `Error`;
-              outGutter.classList.remove('flex', 'items-center', 'justify-end', 'gap-1');
-           }
-           if (streamingContent) {
-              streamingContent.innerHTML = `<span class="text-error">Execution Failed.</span>`;
-              streamingContent.classList.remove('animate-pulse');
-           }
-        } finally {
-           executingCells.delete(logId);
-        }
-      } else {
-        let fullText = "";
-        api.streamQuestion(activeId!, question, logId, null,
-          (token) => {
-            fullText += token;
-            if (streamingContent) {
-              streamingContent.innerHTML = renderMarkdown(fullText) + '<span class="inline-block w-1 h-4 bg-primary animate-pulse ml-1"></span>';
-            }
-          },
-          async () => {
-            executingCells.delete(logId);
-            if (streamingContent) {
-              streamingContent.innerHTML = renderMarkdown(fullText);
-              streamingContent.classList.remove('animate-pulse');
-            }
-            if (outGutter) {
-              outGutter.textContent = `Out [${idx}]:`;
-              outGutter.classList.remove('flex', 'items-center', 'justify-end', 'gap-1');
-            }
-            
-            const newLogs = await api.listLogs(activeId!);
-            lastLogsJson = '';
-            store.setLogs(newLogs);
-          },
-          (err) => {
-            executingCells.delete(logId);
-            console.error('Rerun error', err);
-            if (outGutter) {
-               outGutter.textContent = `Out [${idx}]:`;
-               outGutter.classList.remove('flex', 'items-center', 'justify-end', 'gap-1');
-            }
-          }
-        );
-      }
-    }
-
-    // Run Static Cell directly
-    const runStaticBtn = target.closest('.btn-run-static-cell') as HTMLButtonElement;
-    if (runStaticBtn && activeId) {
-      const idx = runStaticBtn.getAttribute('data-cell-index');
-      const logId = runStaticBtn.getAttribute('data-log-id');
-      const toolName = runStaticBtn.getAttribute('data-tool');
-      if (!idx || !logId || !toolName) return;
-      
-      const txt = container.querySelector(`#cell-input-${idx}`) as HTMLTextAreaElement;
-      const staticView = container.querySelector(`#cell-static-${idx}`) as HTMLElement;
-      if (!staticView) return;
-      
-      const question = txt ? txt.value : staticView.innerText.trim();
-      if (!question) return;
-      
-      runStaticBtn.disabled = true;
-      
-      const cellBody = runStaticBtn.closest('.glass-panel.rounded-3xl') as HTMLElement;
-      if (cellBody) {
-        await runCell(idx, logId, toolName, question, cellBody);
-      }
-      return;
-    }
-
-    // "First Cell" welcome screen chooser
-    const firstCellBtn = target.closest('.btn-first-cell') as HTMLElement;
-    if (firstCellBtn) {
-      const type = parseCellType(firstCellBtn.getAttribute('data-type'));
-      if (!type) return;
-      const cellContainer = container.querySelector('#cell-container');
-      if (!cellContainer) return;
-
-      // Remove the welcome screen
-      const welcome = cellContainer.querySelector('#notebook-welcome');
-      welcome?.remove();
-
-      // Create first cell edit block
-      const dummy = document.createElement('div');
-      dummy.className = 'hidden';
-      cellContainer.appendChild(dummy);
-      createCellEditBlock(type, null, dummy);
-      return;
-    }
-    
-    // Toggle Edit Mode
-    const editBtn = target.closest('.btn-edit-cell') as HTMLElement;
-    if (editBtn) {
-      const idx = editBtn.getAttribute('data-cell-index');
-      container.querySelector(`#cell-static-${idx}`)?.classList.add('hidden');
-      container.querySelector(`#cell-edit-${idx}`)?.classList.remove('hidden');
-      container.querySelector(`#cell-divider-${idx}`)?.classList.add('opacity-0');
-      
-      const txt = container.querySelector(`#cell-input-${idx}`) as HTMLTextAreaElement;
-      if (txt) {
-         setTimeout(() => {
-            txt.style.height = 'auto';
-            txt.style.height = txt.scrollHeight + 'px';
-            txt.focus();
-         }, 0);
-      }
-      return;
-    }
-    
-    // Delete Cell
-    const deleteBtn = target.closest('.btn-delete-cell') as HTMLElement;
-    if (deleteBtn) {
-      const logId = deleteBtn.getAttribute('data-log-id');
-      if (logId && confirm('Are you sure you want to delete this cell?')) {
-        try {
-          await api.deleteLog(logId);
-          const newLogs = await api.listLogs(activeId!);
-          lastLogsJson = '';
-          store.setLogs(newLogs);
-        } catch (e) {
-          console.error('Failed to delete cell', e);
-        }
-      }
-      return;
-    }
-
-    // Cancel Edit Mode
-    const cancelBtn = target.closest('.btn-cancel-edit') as HTMLElement;
-    if (cancelBtn) {
-      const idx = cancelBtn.getAttribute('data-cell-index');
-      container.querySelector(`#cell-static-${idx}`)?.classList.remove('hidden');
-      container.querySelector(`#cell-edit-${idx}`)?.classList.add('hidden');
-      container.querySelector(`#cell-divider-${idx}`)?.classList.remove('opacity-0');
-      return;
-    }
-    
-    // Rerun Cell
-    const rerunBtn = target.closest('.btn-rerun-cell') as HTMLButtonElement;
-    if (rerunBtn && activeId) {
-      const idx = rerunBtn.getAttribute('data-cell-index');
-      const logId = rerunBtn.getAttribute('data-log-id');
-      const txt = container.querySelector(`#cell-input-${idx}`) as HTMLTextAreaElement;
-      if (!txt || !logId || !idx) return;
-      
-      const question = txt.value;
-      if (!question) return;
-      
-      rerunBtn.disabled = true;
-
-      const toolName = rerunBtn.getAttribute('data-tool');
-      if (toolName === 'markdown') {
-        rerunBtn.innerHTML = `<span class="material-symbols-outlined text-[14px] animate-spin">progress_activity</span>`;
-        try {
-           await api.executeMarkdown(activeId, question, logId, null);
-           
-           container.querySelector(`#cell-edit-${idx}`)?.classList.add('hidden');
-           container.querySelector(`#cell-static-${idx}`)?.classList.remove('hidden');
-           container.querySelector(`#cell-divider-${idx}`)?.classList.remove('opacity-0');
-
-           const newLogs = await api.listLogs(activeId);
-           lastLogsJson = '';
-           store.setLogs(newLogs);
-        } catch (e) {
-           console.error(e);
-           rerunBtn.disabled = false;
-           rerunBtn.innerHTML = `<span class="material-symbols-outlined text-[14px]">done</span>`;
-        }
-        return;
-      }
-      
-      const cellBody = rerunBtn.closest('.glass-panel.rounded-3xl') as HTMLElement;
-      if (cellBody) {
-        await runCell(idx, logId, toolName || 'chat', question, cellBody);
-      }
-      return;
-    }
-
-    // Insert New Unexecuted Cell
-    const insertBtn = target.closest('.btn-insert-cell') as HTMLElement;
-    if (insertBtn) {
-      const type = parseCellType(insertBtn.getAttribute('data-type'));
-      const afterLogId = parseAfterLogId(insertBtn.getAttribute('data-after'));
-      const toolbarNode = insertBtn.closest('.group\\/toolbar') as HTMLElement;
-      
-      if (!toolbarNode || !type || !afterLogId) return;
-
-      createCellEditBlock(type, afterLogId, toolbarNode, toolbarNode);
-    }
-
-    // Execute Newly Inserted Cell
-    const runNewBtn = target.closest('.btn-execute-new-cell') as HTMLButtonElement;
-    if (runNewBtn && activeId) {
-       const type = runNewBtn.getAttribute('data-type');
-       const rawAfter = runNewBtn.getAttribute('data-after');
-       // '__first__' is a sentinel meaning "no anchor — just append"
-       const afterLogId = (rawAfter && rawAfter !== '__first__') ? rawAfter : null;
-       const rawTempId = runNewBtn.getAttribute('data-temp-id');
-       const tempId = rawTempId && /^[A-Za-z0-9_-]+$/.test(rawTempId) ? rawTempId : null;
-       
-       if (!type || !tempId) return;
-       
-       const txt = container.querySelector(`#cell-input-${tempId}`) as HTMLTextAreaElement;
-       const question = txt?.value;
-       if (!question) return;
-
-       runNewBtn.disabled = true;
-
-       const cellBody = runNewBtn.closest('.glass-panel.rounded-3xl');
-       if (!cellBody) return;
-
-       if (type === 'markdown') {
-          runNewBtn.innerHTML = `<span class="material-symbols-outlined text-[14px] animate-spin">progress_activity</span>`;
-          try {
-             await api.executeMarkdown(activeId, question, null, afterLogId);
-             
-             const newCellBlock = runNewBtn.closest('.new-cell-block');
-             newCellBlock?.remove();
-
-             const newLogs = await api.listLogs(activeId);
-             lastLogsJson = '';
-             store.setLogs(newLogs);
-          } catch (e) {
-             console.error(e);
-             runNewBtn.disabled = false;
-             runNewBtn.innerHTML = `<span class="material-symbols-outlined text-[14px]">done</span>`;
-          }
-          return;
-       }
-
-       // Transition UI to executing state
-       const editView = container.querySelector(`#cell-edit-${tempId}`);
-       if (editView) {
-          const staticDiv = document.createElement('div');
-          staticDiv.className = 'flex-1 font-mono text-sm text-primary-fixed-dim bg-surface-container-lowest/80 border border-outline-variant/10 rounded-xl p-3.5 shadow-inner overflow-x-auto';
-          staticDiv.textContent = question;
-          editView.replaceWith(staticDiv);
-       }
-
-       // Append Output Gutter
-       const outputDiv = document.createElement('div');
-       outputDiv.className = 'flex items-start gap-4 mt-6 border-t border-outline-variant/10 pt-6';
-       outputDiv.innerHTML = `
-          <div class="font-mono text-xs font-bold text-secondary/50 pt-1 select-none w-14 text-right shrink-0 flex items-center justify-end gap-1 out-label">
-             <span class="material-symbols-outlined text-[10px] animate-spin" data-icon="progress_activity">progress_activity</span>
-             <span>Out [*]:</span>
-          </div>
-          <div class="flex-1 min-w-0" id="out-body-${tempId}">
-             <div class="prose prose-invert max-w-none text-on-surface-variant leading-relaxed font-body-lg animate-pulse" id="streaming-content-${tempId}">
-                <span class="inline-block w-1 h-4 bg-primary animate-pulse"></span>
-             </div>
-          </div>
-       `;
-       cellBody.appendChild(outputDiv);
-
-       const outGutter = outputDiv.querySelector('.out-label') as HTMLElement;
-       const streamingContent = outputDiv.querySelector(`#streaming-content-${tempId}`);
-
-       if (type === 'chat') {
-          let fullText = "";
-          api.streamQuestion(activeId, question, null, afterLogId,
-             (token) => {
-                fullText += token;
-                if (streamingContent) streamingContent.innerHTML = renderMarkdown(fullText) + '<span class="inline-block w-1 h-4 bg-primary animate-pulse ml-1"></span>';
-             },
-             async () => {
-                const newCellBlock = runNewBtn.closest('.new-cell-block');
-                newCellBlock?.remove();
-
-                const newLogs = await api.listLogs(activeId!);
-                lastLogsJson = '';
-                store.setLogs(newLogs);
-             },
-             (err) => {
-                console.error(err);
-                if (outGutter) {
-                   outGutter.textContent = 'Error';
-                   outGutter.classList.remove('flex', 'items-center', 'justify-end', 'gap-1');
-                }
-                if (streamingContent) {
-                   streamingContent.innerHTML += '<br><span class="text-error">Execution Failed.</span>';
-                   streamingContent.classList.remove('animate-pulse');
-                }
-             }
-          );
-       } else if (type === 'sql' || type === 'python') {
-          try {
-             if (type === 'sql') {
-                await api.executeSql(activeId, question, null, afterLogId);
-             } else {
-                await api.executePython(activeId, question, null, afterLogId);
-             }
-
-             const newCellBlock = runNewBtn.closest('.new-cell-block');
-             newCellBlock?.remove();
-
-             const newLogs = await api.listLogs(activeId);
-             lastLogsJson = '';
-             store.setLogs(newLogs);
-          } catch (e) {
-             console.error(e);
-             if (outGutter) {
-                outGutter.textContent = `Error`;
-             }
-             if (streamingContent) {
-                streamingContent.innerHTML = `<span class="text-error">Execution Failed.</span>`;
-                streamingContent.classList.remove('animate-pulse');
-             }
-          }
-       }
-    }
-
-     // Data Source Preview Modal trigger
-     const previewBtn = target.closest('.btn-preview-data') as HTMLElement;
-     if (previewBtn) {
-       const fullTable = previewBtn.getAttribute('data-table');
-       const filename = previewBtn.getAttribute('data-filename');
-       if (fullTable) {
-         showDataPreviewModal(fullTable, filename || 'Data Source');
-       }
-     }
-
-     // Insert Table Name trigger
-     const insertTableBtn = target.closest('.btn-insert-table') as HTMLElement;
-     if (insertTableBtn) {
-       const fullTable = insertTableBtn.getAttribute('data-table');
-       if (fullTable) {
-         let activeTextarea: HTMLTextAreaElement | null = null;
-         
-         // 1. Try currently focused
-         if (document.activeElement && document.activeElement.tagName === 'TEXTAREA' && container.contains(document.activeElement)) {
-           activeTextarea = document.activeElement as HTMLTextAreaElement;
-         } 
-         // 2. Try the new cell block
-         else if (container.querySelector('.new-cell-block textarea')) {
-           activeTextarea = container.querySelector('.new-cell-block textarea') as HTMLTextAreaElement;
-         }
-         // 3. Try any open edit view
-         else {
-           const editViews = container.querySelectorAll('.cell-edit-view:not(.hidden) textarea');
-           if (editViews.length > 0) {
-             activeTextarea = editViews[0] as HTMLTextAreaElement;
-           }
-         }
-
-         if (activeTextarea) {
-           const start = activeTextarea.selectionStart;
-           const end = activeTextarea.selectionEnd;
-           const val = activeTextarea.value;
-           activeTextarea.value = val.substring(0, start) + fullTable + val.substring(end);
-           activeTextarea.selectionStart = activeTextarea.selectionEnd = start + fullTable.length;
-           activeTextarea.dispatchEvent(new Event('input', { bubbles: true }));
-           activeTextarea.focus();
-         }
-       }
-     }
-  });
-}
-
-async function showDataPreviewModal(fullTableName: string, filename: string) {
-  let modal = document.getElementById('data-preview-modal');
-  if (!modal) {
-    modal = document.createElement('div');
-    modal.id = 'data-preview-modal';
-    modal.className = 'fixed inset-0 z-[100] flex items-center justify-center bg-background/80 backdrop-blur-md opacity-0 pointer-events-none transition-opacity duration-300';
-    modal.innerHTML = `
-      <div class="glass-panel p-6 rounded-3xl w-11/12 max-w-5xl max-h-[85vh] flex flex-col shadow-2xl shadow-primary/20 bg-surface-container-low/90 border-outline-variant/20 scale-95 transition-transform duration-300" id="data-preview-content">
-        <div class="flex justify-between items-center mb-4 shrink-0">
-          <div class="flex items-center gap-3 min-w-0">
-            <div class="w-12 h-12 rounded-2xl bg-primary/20 flex items-center justify-center text-primary shrink-0">
-              <span class="material-symbols-outlined">database</span>
-            </div>
-            <div class="min-w-0">
-              <h3 class="text-xl font-headline-sm text-white truncate max-w-xl" id="preview-title">Data Preview</h3>
-              <p class="text-[11px] text-primary/70 font-mono tracking-widest uppercase mt-0.5" id="preview-subtitle">Loading...</p>
-            </div>
-          </div>
-          <button class="w-10 h-10 rounded-full bg-surface-container-highest flex items-center justify-center hover:bg-error/20 hover:text-error transition-colors shrink-0" id="btn-close-preview">
-            <span class="material-symbols-outlined">close</span>
-          </button>
-        </div>
-        <div class="flex-1 min-h-0 overflow-auto custom-scrollbar bg-surface-container-lowest/50 rounded-2xl border border-outline-variant/10" id="preview-table-container">
-           <div class="flex items-center justify-center h-40">
-             <span class="material-symbols-outlined animate-spin text-primary text-4xl">progress_activity</span>
-           </div>
-        </div>
-      </div>
-    `;
-    document.body.appendChild(modal);
-    
-    modal.querySelector('#btn-close-preview')?.addEventListener('click', () => {
-      modal?.classList.add('opacity-0', 'pointer-events-none');
-      modal?.querySelector('#data-preview-content')?.classList.replace('scale-100', 'scale-95');
-    });
-    modal.addEventListener('click', (e) => {
-      if (e.target === modal) {
-        modal?.classList.add('opacity-0', 'pointer-events-none');
-        modal?.querySelector('#data-preview-content')?.classList.replace('scale-100', 'scale-95');
-      }
-    });
-  }
-  
-  const titleEl = modal.querySelector('#preview-title');
-  const subtitleEl = modal.querySelector('#preview-subtitle');
-  const tableContainer = modal.querySelector('#preview-table-container');
-  
-  if (titleEl) titleEl.textContent = filename || 'Data Preview';
-  if (subtitleEl) subtitleEl.textContent = `Table: ${fullTableName}`;
-  if (tableContainer) {
-    tableContainer.innerHTML = `<div class="flex items-center justify-center h-64"><span class="material-symbols-outlined animate-spin text-primary text-4xl shadow-primary/20 shadow-lg rounded-full">progress_activity</span></div>`;
-  }
-  
-  modal.classList.remove('opacity-0', 'pointer-events-none');
-  modal.querySelector('#data-preview-content')?.classList.replace('scale-95', 'scale-100');
-  
-  try {
-    const data = await api.getTablePreview(fullTableName);
-    
-    if (!data || data.length === 0) {
-      if (tableContainer) tableContainer.innerHTML = `<div class="flex flex-col items-center justify-center h-64 text-outline"><span class="material-symbols-outlined text-4xl mb-2 opacity-50">draft</span><p>No rows found</p></div>`;
-      return;
-    }
-    
-    const keys = Object.keys(data[0]);
-    let html = `
-      <table class="w-full text-left border-collapse text-sm font-mono">
-        <thead class="bg-surface-container-highest text-xs uppercase tracking-wider text-primary sticky top-0 shadow-sm z-10">
-          <tr>${keys.map(k => `<th class="px-6 py-4 font-medium whitespace-nowrap border-b border-primary/20">${k}</th>`).join('')}</tr>
-        </thead>
-        <tbody class="divide-y divide-outline-variant/10">`;
-        
-    data.forEach((row: any) => {
-      html += `<tr class="hover:bg-surface-container-low/70 transition-colors">
-        ${keys.map(k => {
-           let val = row[k];
-           if (val === null || val === undefined) return '<td class="px-6 py-3"><span class="text-outline/40 italic text-[11px]">null</span></td>';
-           if (typeof val === 'object') val = JSON.stringify(val);
-           return `<td class="px-6 py-3 text-on-surface-variant truncate max-w-sm" title="${val}">${val}</td>`;
-        }).join('')}
-      </tr>`;
-    });
-    
-    html += `</tbody></table>`;
-    if (tableContainer) tableContainer.innerHTML = html;
-    
-  } catch (err) {
-    if (tableContainer) tableContainer.innerHTML = `<div class="flex flex-col items-center justify-center h-64 text-error"><span class="material-symbols-outlined text-4xl mb-2 opacity-80">error</span><p>Failed to load data preview.</p></div>`;
-  }
-}
